added docstrings, relocated few files
