+++ frontend/src/config/apollo_cache.ts
@@ -2,7 +2,6 @@ import { ApolloCache } from "@apollo/client";
 import { MY_PRODUCTS_QUERY } from "../graphql/product/queries";
 import { Product } from "../constants/types/Product";
 export function updateOnNew(cache: ApolloCache<unknown>, product: Product) {
-    console.log(product);
     const { getMyProducts } = cache.readQuery({ query: MY_PRODUCTS_QUERY });
     cache.writeQuery({
         query: MY_PRODUCTS_QUERY,
