+++ frontend/src/pages/Homepage.tsx
@@ -1,5 +1,5 @@
-import { Box, Button } from "@mui/joy";
 import { Link } from "react-router-dom";
+import { Box, Button } from "@mui/joy";
 
 export function Homepage() {
     return (
