+++ frontend/src/pages/components/containers/index.ts
@@ -1,0 +1,6 @@
+import FormContainer from "./FormContainer";
+import InputFeedbackContainer from "./InputFeedbackContainer";
+import PositionElement from "./PositionElement";
+import RequestStateWrapper from "./RequestStateWrapper";
+
+export { FormContainer, InputFeedbackContainer, PositionElement, RequestStateWrapper };
