+++ frontend/src/pages/components/products/ViewProducts.tsx
@@ -1,8 +1,7 @@
 import { Box, Typography } from "@mui/joy";
-import { Category, Product, RentDuration } from "../../../constants/types/Product";
+import { Product } from "../../../constants/types/Product";
 import ProductCard from "./ProductCard";
 import PositionElement from "../containers/PositionElement";
-import { getEnumValues } from "../../../utils/helper";
 
 interface Props {
     products: Product[];
@@ -11,6 +10,16 @@ interface Props {
     forMyPage?: boolean;
 }
 
+/**
+ * View The Page title for my products and all products.
+ * Deligate the product rendering task to Product card component
+ *
+ * @param {Readonly<Props>} products - the list of products to display
+ * @param {string} title - the title to display above the list of products
+ * @param {string} linkPrefix - the prefix to add to each product link (differentiate between my products and all products page)
+ * @param {boolean} forMyPage - flag indicating if the products are for the user's page
+ * @return {JSX.Element} the JSX element representing the list of products
+ */
 export default function ViewProducts({ products, title, linkPrefix, forMyPage = false }: Readonly<Props>) {
     if (!products || products.length === 0) {
         return <Typography level="title-lg">No Products Found</Typography>;
@@ -24,11 +33,7 @@ export default function ViewProducts({ products, title, linkPrefix, forMyPage = 
                         productLink={linkPrefix + product.id}
                         key={product.id}
                         showDelete={forMyPage}
-                        product={{
-                            ...product,
-                            categories: getEnumValues(Category, product.category),
-                            rentDuration: RentDuration[product.rentDuration],
-                        }}
+                        product={product}
                     ></ProductCard>
                 ))}
             </Box>
