+++ frontend/src/pages/components/products/transaction/ProductFooter.tsx
@@ -1,8 +1,8 @@
-import { Product } from "../../../../constants/types/Product";
-import RequestStateWrapper from "../../containers/RequestStateWrapper";
 import { Typography } from "@mui/joy";
-import PositionElement from "../../containers/PositionElement";
 import { useEffect } from "react";
+
+import { Product } from "../../../../constants/types/Product";
+import { PositionElement, RequestStateWrapper } from "../../containers";
 
 interface ProductFooterProps {
     product?: Product;
@@ -12,6 +12,16 @@ interface ProductFooterProps {
     loading: boolean;
 }
 
+/**
+ * Renders Buttons/State of Product with appropriate message.
+ *
+ * @param {Readonly<ProductFooterProps>} product - Product to retrieve ownerId and get if it is loaded
+ * @param {ReactNode} children - The child elements to be rendered within the footer (if it is buyable)
+ * @param {string} userId - logged in user id (useAuth creates some issue with multiple re-rendering)
+ * @param {TransactionHistory[]} transactionHistory - The transaction history associated with the product.
+ * @param {boolean} loading - A boolean indicating transaction history loading state.
+ * @return {ReactNode} The rendered product footer component.
+ */
 export default function ProductFooter({
     product,
     children,
@@ -40,8 +50,8 @@ export default function ProductFooter({
                             {isBoughtByUser
                                 ? "You Bought this Product"
                                 : isSoldByUser
-                                  ? "You Sold this Product"
-                                  : "Product Sold"}
+                                ? "You Sold this Product"
+                                : "Product Sold"}
                         </b>
                     </Typography>
                 )}
