+++ frontend/src/pages/components/resuable/form_input/InputFeedback.tsx
@@ -1,25 +1,8 @@
 import { Input } from "@mui/joy";
 import { useFormContext } from "react-hook-form";
-import { InputFeedbackContainer } from "../../containers/InputFeedbackContainer";
 import { SxProps } from "@mui/joy/styles/types";
 
-// interface InputFeedBackProps {
-//     children: React.ReactNode;
-//     getInputFeedback(value: string): string;
-//     value: string;
-// }
-
-// function InputFeedBack(props: InputFeedBackProps) {
-//     const { children, getInputFeedback, value } = props;
-//     return (
-//         <Stack spacing={0.5}>
-//             {children}
-//             <Typography level="body-xs" sx={{ alignSelf: "flex-end" }}>
-//                 <b>{getInputFeedback(value)}</b>
-//             </Typography>
-//         </Stack>
-//     );
-// }
+import { InputFeedbackContainer } from "../../containers";
 
 interface FormInputProps {
     id: string;
@@ -29,7 +12,7 @@ interface FormInputProps {
     inputStyles?: SxProps;
 }
 
-export function FormInput(props: FormInputProps) {
+export default function FormInput(props: FormInputProps) {
     const { id, placeholder, type, styles, inputStyles } = props;
     const {
         register,
