+++ frontend/src/pages/components/resuable/form_input/PasswordField.tsx
@@ -4,13 +4,20 @@ import Visibility from "@mui/icons-material/Visibility";
 import VisibilityOff from "@mui/icons-material/VisibilityOff";
 import { useState } from "react";
 import { useFormContext } from "react-hook-form";
-import { InputFeedbackContainer } from "../../containers/InputFeedbackContainer";
+import { InputFeedbackContainer } from "../../containers";
 
 interface Props {
     id: string;
     placeholder?: string;
 }
 
+/**
+ * Generates a password input form field with toggle visibility functionality.
+ *
+ * @param {Props} id - The unique identifier for the input field
+ * @param {Props} placeholder - The placeholder text for the input field (default is "Password")
+ * @return {JSX.Element} The password input form field component
+ */
 export default function FormPasswordInput({ id, placeholder = "Password" }: Props) {
     const [showPassword, setShowPassword] = useState(false);
