+++ frontend/src/pages/layout/Mainlayout.tsx
@@ -1,6 +1,12 @@
 import { Outlet } from "react-router-dom";
 import AppBar from "../components/Appbar";
 
+/**
+ * Adds AppBar and to component.
+ * Renders the main layout of the application.
+ *
+ * @return {JSX.Element} The main layout component
+ */
 export default function MainLayout() {
     return (
         <>
