+++ frontend/src/pages/layout/RequireAuth.tsx
@@ -1,6 +1,12 @@
 import { Navigate, Outlet, useLocation } from "react-router-dom";
 import useAuth from "../../config/hooks/useAuth";
 
+/**
+ * Component that requires authentication before rendering the content.
+ * Wrapper For the protected routes.
+ *
+ * @return {JSX.Element} The rendered JSX based on the authentication status.
+ */
 export default function RequireAuth() {
     const {
         userData: { id },
